requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain & LangGraph
langchain==0.1.6
langchain-google-genai==0.0.6
langchain-community==0.0.20
langgraph==0.0.20

# Vector Store
chromadb==0.4.22

# Environment & Utils
python-dotenv==1.0.0
